Revert "Feat/déplacement ordonanceur (#5)"
This reverts commit 3978aaa31631bc96a3adfbb0496e7f952d8d15c2.
